chore: fix invisible meter-bar border + prune dead code flagged by CI warnings (#47)
- StatusPanel: drawRect on an int rectangle truncated the 0.5f line
  thickness to 0, so the fill-bar border never drew; use the float
  overload so the intended hairline actually renders.
- MainComponent: kTabFade was defined in the .cpp but the real default
  lived as a duplicated 0.30 literal in the header; move both fade
  constants into the class so switchTab's default references the single
  source of truth (-Wunused-const-variable).
- SpectralProcessor: drop write-only fftOrderVal field
  (-Wunused-private-field); the FFT already owns the order.
- Main: tray onQuit lambda captured this for a static call
  (-Wunused-lambda-capture).

+++ Source/UI/StatusPanel.cpp
@@ -418,7 +418,7 @@ namespace dcr
                 g.setColour (valueCol);
                 g.fillRect (barArea.withWidth (filledW));
                 g.setColour (juce::Colour::fromRGB (40, 40, 48));
-                g.drawRect (barArea, 0.5f);
+                g.drawRect (barArea.toFloat(), 0.5f); // int overload would truncate 0.5 -> 0 (no border)
                 inner.removeFromTop (2);
             }
